requirements.txt
azure-ai-translation-text == 1.0.0b1
azure-storage-blob
flask